chore: Update npm dependencies and refactor import paths
This commit updates the npm dependencies in the package.json file, adding "jwt-decode" and "universal-cookie" packages. Additionally, it refactors import paths for the "IKeypair" type and the "Environment Mutation" to improve code organization and maintainability. These changes have no functional impact and are purely for dependency management and import structure.