Add Stellar transaction signing in Invocation Page
Updated InvocationPage to support transaction signing functionality to enhance app's Stellar network integration. The change includes leveraging the simple-stellar-signer-api to sign transactions, and state and effect hooks to manage the signed Transaction XDR. Also adapted wallet usage within the page by incorporating useAuthProvider hook. This allows validating the existence of public keys in the wallet before signing the transaction.

As an impact, this results in a more secure, versatile and user-friendly invocation stage in terms of transaction handling on the Stellar network.