Added prepare invocation functionality in Contract Input
In reaction to new requirement, the additional functionality to prepare invocations has been integrated into Contract Input component. This change introduces a 'PREPARE' button which becomes available when the contract is loaded and no wallet public key is provided or when the contract is signed. This improves usability and prepares the software for the upcoming contract invocation feature.