Removed Authorization tab from TabsContainer component
In this update, we've streamlined the TabsContainer component by removing the Authorization tab. This is part of an effort to declutter the UI and improve user experience. The removed code includes not only the tab itself but also related handling functions and tooltips. As a consequence, users will no longer see authorization related functionality in this specific component.