Removed AUTO_DETECT from NETWORK Enum
The AUTO_DETECT option has been removed from the NETWORK enum in the soroban scope. This change was necessary in order to streamline the network selection process, and to minimize the potential for network detection errors. Care should be taken when setting the network type going forward, as automatic detection is no longer available.