Refactor wallet handling in AuthorizationTab
Updates the useAuthProvider hook to use the correct function name for disconnecting wallet, 'disconnectWallet', instead of the previous 'setDisconnectWallet'. Also extends the responsiveness of the component to changes in the 'network' prop. The public key retrieval now accounts for potential variation in key location depending on the network type, increasing robustness and reducing the risk of key fetching errors.