Simplified asynchronous mutation in CollectionPlaceholder
Removed unnecessary async-await syntax in the onSubmit handler of the CollectionPlaceholder component. The change was made because the 'mutate' function doesn't need to be awaited as it doesn't return anything that we need to wait for. This simplifies the code and might help slightly with performance. No side effects expected from this change as the function after mutate is not relying on its result.