Implemented message handler in signer API module
Introduced a new method `messageHandler` in the `signerApi` module. This function will handle sign and connect messages for the popup window, thus improving message management.

It listens to 'message' events, and has handlers set up for 'READY' and 'CANCEL' events types. On 'READY' it initiates the postMessage with the given config and on 'CANCEL' it smoothly interrupts the process.

This approach efficiently deals with window timeout scenarios, enhancing the application's resilience against unexpected failures.